Refactor rename components folder to features
