+++ app/src/features/NPCGenerator.tsx
